pyproject.toml
[tool.poetry]
name = "djangoproject"
version = "0.1.0"
description = ""
authors = ["saros1249 <saros1249@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
psycopg2 = "^2.9.5"
pillow = "^9.3.0"
sqlparse = "^0.4.3"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
